exec: put address space dispatch under RCU critical section
With this change, address space dispatch can be moved outside the
BQL.  The actual I/O still has to happen within the lock.

The complicated thing here is determining the boundary of
TCG's read-side critical sections, but the previous patches set it
up nicely.  CPU_INTERRUPT_TLBFLUSH will be triggered (and force
an exit of the current translation block) as soon as RCU-protected
data changes.  Not long after, when the interrupt is processed, the
CPUArchState will be updated with the new data structure and the old
ones will not be used anymore.  This means that CPU_INTERRUPT_TLBFLUSH
can be used as the boundary of an RCU read-side critical section.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>