pci-assign: move memory_region_destroy of BARs to instance_finalize
free_assigned_device is also doing memory_region_del_subregion for the
"real_iomem" regions.  This is not a problem, because these regions are
always nested in a container region; these container regions are passed
to pci_register_bar and are unreachable after pci_unregister_io_regions
returns, hence the same thing is true of "real_iomem" regions.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>