tcg-ppc: use carry-based sequences for unsigned setcond
For constants we could also use a trick similar to the one in the next
commit:

    x > y (unsigned)         sign bit of x | (y - x), if y sign bit is 0
    x > y (unsigned)         sign bit of x & (y - x), if y sign bit is 1

x <= y can then be computed with NAND and NOR, while x < y and x >= y are
obtained by adjusting y.  However, the sequences produced by this patch
are three instructions long, so leave this for later.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>