optimize: optimize deposit to a previous movi destination
With x86 a common pattern is to zero a register and then do a setCC
operation into it.  In this case, a TCG setcond operation is enough
and there is no need for a deposit operation.  With this patch, the
optimizer notices this and makes the previous store of zero dead.
This gives a ~0.5% improvement for various i386-on-PPC benchmarks,
and a ~1% reduction in executed instructions.

The optimization hits in other cases, such as moving a constant
to %bx where the high bits of %ebx are known to be zero.  This
requires the special handling of deposit operations that become
a movi.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>